requirements.txt
spacy
python-telegram-bot
cedict-utils
